Convert Naming options to TypeScript
(cherry picked from commit 8b64f873f47e732cf318c2874695f087261c948d)
