+++ frontend/src/Settings/MediaManagement/Naming/NamingOption.css.d.ts
@@ -8,6 +8,7 @@ interface CssExports {
   'lower': string;
   'option': string;
   'small': string;
+  'title': string;
   'token': string;
   'upper': string;
 }
