+++ frontend/src/Settings/MediaManagement/Naming/TokenCase.ts
@@ -1,0 +1,3 @@
+type TokenCase = 'title' | 'lower' | 'upper';
+
+export default TokenCase;
