+++ frontend/src/Settings/MediaManagement/Naming/TokenSeparator.ts
@@ -1,0 +1,3 @@
+type TokenSeparator = ' ' | '.' | '_' | '-';
+
+export default TokenSeparator;
