+++ frontend/src/typings/Settings/NamingExample.ts
@@ -1,0 +1,7 @@
+export default interface NamingExample {
+  movieExample: string;
+  movieFolderExample: string;
+  sceneExample: string;
+  sceneFolderExample: string;
+  sceneImportFolderExample: string;
+}
