+++ src/NzbDrone.Core/Localization/Core/en.json
@@ -681,6 +681,7 @@
   "FocusSearchBox": "Focus Search Box",
   "Folder": "Folder",
   "FolderMoveRenameWarning": "This will also rename the movie folder per the movie folder format in settings.",
+  "FolderNameTokens": "Folder Name Tokens",
   "Folders": "Folders",
   "FollowPerson": "Follow Person",
   "Forecast": "Forecast",
